oz-django/pyproject.toml
[tool.poetry]
name = "oz-django"
version = "0.1.0"
description = ""
authors = ["Samiiz <skboss159@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
